+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Compact Log Event Format (CLEF) formatter.
 - Add functions "delay_input", "delay_output", "to_base64", "from_base64".
 - Add "call" command.
+- Initial implementation of inputs with references to other columns.
 
 ### Fixed